# [Perf] Optimize Qwen2/2.5-VL ViT tensor generating performance