--- a/requirements/rocm.txt
+++ b/requirements/rocm.txt
@@ -1,9 +1,6 @@
 # Common dependencies
 -r common.txt
 
-numba == 0.60.0; python_version == '3.9' # v0.61 doesn't support Python 3.9. Required for N-gram speculative decoding
-numba == 0.61.2; python_version > '3.9'
-
 # Dependencies for AMD GPUs
 awscli
 boto3