… tests/model_executor/test_qwen2_vl_rot_pos.py