@NickLucche 2025-04-15
Nice work!
I wonder whether it would've been easier to land this contribution in two separate PRs, one with the CUDA implementation and a follow-up with the numba one for CPU.