@imkero 2025-04-18
> Nice work! I wonder whether it would've been easier to land this contribution in two separate PRs, one with the CUDA implementation and a follow-up with the numba one for CPU.

I tried all possible solutions I know (numba only, torch + numba, triton) for this PR to find the most efficient approach, so I would like to bring the most efficient solution I found for both CPU and CUDA backend here. And it's quite simple (just adding a dispatch function in front of them)